+++ src/SimpleCrawler.Js.V8/V8ModuleLoader.cs
@@ -1,4 +1,3 @@
-using SimpleCrawler.Core.Helpers;
 using SimpleCrawler.Js.Abstractions;
 using Microsoft.ClearScript;
 using Microsoft.ClearScript.JavaScript;
@@ -56,10 +55,7 @@ internal sealed class V8ModuleLoader : DocumentLoader
 
     private Uri Resolve(DocumentInfo? sourceInfo, string specifier)
     {
-        if (UriHelper.TryCreateHttpAbsolute(specifier, out var absolute))
-            return absolute;
-
         var referrer = sourceInfo?.Uri is { IsAbsoluteUri: true } source ? source : _baseUri;
-        return new Uri(referrer, specifier);
+        return ModuleSpecifier.Resolve(specifier, referrer, _fetcher.ImportMap);
     }
 }